Separate navigation Tabs to it's own .tsx file in ui
